feat(symbols/package): Added endpoint for downloading latest symbols package

+++ hexproof/utils/archive.py
@@ -8,28 +8,37 @@ import zipfile
 from pathlib import Path
 
 
-def unpack_tar_gz(path: Path, remove: bool = True) -> None:
-    """Unpacks a `.tar.gz` archive.
+def unpack_zip(path: Path, remove: bool = True) -> None:
+    """Unpack target 'zip' archive.
 
     Args:
-        path: Path to the archive file.
-        remove: Whether to remove the zip after unpacking.
+        path: Path to the archive.
+        remove: Whether to remove the archive after unpacking.
+
+    Raises:
+        FileNotFoundError: If archive couldn't be located.
     """
-    with tarfile.open(path, 'r:gz') as tar:
-        tar.extractall(path=path.parent)
+    if not path.is_file():
+        raise FileNotFoundError(f'Archive not found: {str(path)}')
+    with zipfile.ZipFile(path, 'r') as zip_ref:
+        zip_ref.extractall(path=path.parent)
     if remove:
         os.remove(path)
 
 
-def unpack_zip(path: Path, remove: bool = True) -> None:
-    """Unpacks a `.zip` archive.
+def unpack_tar_gz(path: Path, remove: bool = True) -> None:
+    """Unpack target `tar.gz` archive.
 
     Args:
-        path: Path to the archive file.
-        remove: Whether to remove the zip after unpacking.
+        path: Path to the archive.
+        remove: Whether to remove the archive after unpacking.
+
+    Raises:
+        FileNotFoundError: If archive couldn't be located.
     """
-    # Unpack the zip file
-    with zipfile.ZipFile(path, 'r') as zip_ref:
-        zip_ref.extractall(path=path.parent)
+    if not path.is_file():
+        raise FileNotFoundError(f'Archive not found: {str(path)}')
+    with tarfile.open(path, 'r:gz') as tar:
+        tar.extractall(path=path.parent)
     if remove:
         os.remove(path)
